fix function ID , add function instance running status update (#29)
* added timestamp to status updates

* feat: added timestamp and requested event fields to StatusUpdate

* feat: added Running event call when a container starts running

* fix: use correct functionID in Leaf

* fix: instantiate correct FUNCTION_ID env var in containers;

this needed a slight change in the containerRuntime interface: now a functionID is needed to start a container, so it can be passed onto it.
Up until this point, the Image Tag was being used as a functionID , but since the KVStore was introduced, we now generate UUIDs for every registered function.

* test: update integration test status update expectations